Optimized requests session management
